+++ camera_system/motion_converter_mock.py
@@ -1,9 +1,9 @@
 """動作変換ができるまでの代替モジュール.
-​
+
 2つの仮想走行体からゲーム動作を求める.
 @author miyashita64
 """
-​
+
 from robot import Robot
 from node import Node, NodeType
 from game_area_info import GameAreaInfo
@@ -15,17 +15,27 @@ from middle_to_middle import MiddleToMiddle
 from middle_to_block import MiddleToBlock
 from block_to_middle import BlockToMiddle
 from block_to_intersection import BlockToIntersection
-from return_to_intersection import ReturnToIntersection
-from return_to_middle import ReturnToMiddle
-from return_to_block import ReturnToBlock
-​
+
+
 class MotionConverterMock():
+    """動作変換を代替するクラス."""
+
     @classmethod
     def coord_to_node(cls, coord) -> Node:
-        return [node for node in GameAreaInfo.node_list if node.coord==coord][0]
-​
+        """指定した座標を持つノードを返す."""
+        return [node for node in GameAreaInfo.node_list if node.coord == coord][0]
+
     @classmethod
     def convert(cls, before: Robot, after: Robot) -> GameMotion:
+        """走行体状態が遷移するために必要なゲーム動作を返す.
+
+        Args:
+            before: 遷移前状態
+            after:  遷移後状態
+
+        Returns:
+            GameMotion: 遷移に必要なゲーム動作
+        """
         before_type = cls.coord_to_node(before.coord).nodeType
         after_type = cls.coord_to_node(after.coord).nodeType
         if before_type == NodeType.CROSS:
